+++ shell.nix
@@ -1,0 +1,4 @@
+{ mkShell, go_1_18, gopls }:
+mkShell rec {
+  buildInputs = [ go gopls_1_18 ];
+}